Fixed project layout + streamlining methodcalls
